Implement routes for sql functions
