+++ backend/account-service/controllers/sqlUserController.js
@@ -1,14 +1,6 @@
 const mysql = require('mysql2/promise');
 require('dotenv').config();
-
-// Create a connection pool
-const pool = mysql.createPool({
-  host: process.env.MYSQL_HOST,
-  user: process.env.MYSQL_USER,
-  password: process.env.MYSQL_PASSWORD,
-  database: process.env.MYSQL_DATABASE
-});
-
+const pool = require('../server')
 
 // Create a user
 // async function createUser(username, password, email) {
@@ -290,3 +282,11 @@ const login = async (req, res) => {
 //       console.error(err);
 //     });
 // })();
+module.exports = {
+  getAllUsers,
+  getSingleUser,
+  createUser,
+  deleteUser,
+  updateUser,
+  login
+}
